Unspecify a version for the required module Microsoft.PowerShell.ConsoleGuiTools
I have tried **numerous** combinations of different ways of formatting this
field vis-a-vis various versions of the module, various setups of the client
installation environment, and various verions of the installer (both the older
PowerShellGet as well as the newer Microsoft.PowerShell.PSResourceGet). I have
also tried manually manipulating the module manifests and module info files
with respect to version specifications and dependency listings. None of it
made any progress.

In the end, no matter what I did, if I added a hashtable entry to RequiredModules
containing 'Microsoft.PowerShell.ConsoleGuiTools' for the ModuleName field and
and any value for the ModuleVersion field, the module manifest failed to validate
(via PowerShellGet\Tets-ModuleManifest) and no useful error messages were given.

Interestingly, the aforementioned formulation of the fully-qualified module name
(i.e. using the ModuleVersion field) **did** successfully validate if *any*
version of the module were already installed locally, even if it were not imported.
Furthermore, the fully-qualified module name **did** successfully validate for
`@{ModuleName='Microsoft.PowerShell.ConsoleGuiTools'}` if the RequiredVersion field
were used *instead of* of the ModuleVersion field, even if the value of the fields
was otherwise the same.

Honestly, I'm not sure what to make of this... the only possible cause that I've
been able to find evidence for is that the the format of the version field of the
module Microsoft.PowerShell.ConsoleGuiTools has changed in a self-incomparable way
between published versions. Specifically, the last 3 published versions (excluding
pre-releases) of Microsoft.PowerShell.ConsoleGuiTools to PSGallery have the following
versions (verbatim): "0.7.6.0", "0.7.5.35", "0.7.4". Considering that the module is
a binary module, that its versions have gone from semver-format to dotnet-format, and
that PSGallery is in the process of moving from exclusively dotnet-format to preferring
semver-format, this might be causing some sort of version validation or comparison bug.
...but there's not much we can do about that, and this is a shaky conjecture anyway.

As such, I settled on the following solution to both automatically (and successfully)
co-install the dependencies of this module (F7History) as well as allow users to
install and concurrently use *different* versions of the dependencies without borking
this module:

- Require the module Microsoft.PowerShell.ConsoleGuiTools **but not** any version spec.

In general, this should still work since Install-Module will default to the latest
(non-pre-release) version of a module required by name but not version. Hopefully, a
minimum required version specifier can be re-introduced in the future if-and-when
Microsoft.PowerShell.ConsoleGuiTools becomes stable (presumably by version 1.0.0).